refactor: Split LSP server into separate endpoint files
- Split server.go into modular endpoint files for better organization
- health.go: Health check endpoint handler
- initialize.go: LSP initialize endpoint handler
- document.go: Document lifecycle endpoints (didOpen, didChange, didClose)
- hover_completion.go: Hover and completion endpoints
- shutdown.go: Shutdown and exit endpoints
- server.go: Cleaned up to contain only core server functionality and method router
- Fixed test failures in hover and completion endpoints
- All tests passing with improved code organization
